Add my_enumerable module
Create enumerable methods:
-all?
-any?
-filter
-Correct linter errors and update file names
