--- enumerable.rb
@@ -1,1 +1,0 @@
-puts 'Hello Our Enumerable'
